+++ flake-ctl-v1/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "flake-ctl"
+version = "2.2.19"
+authors = ["Marcus Schäfer <marcus.schaefer@elektrobit.com>"]
+edition = "2018"
+license = "MIT"
+
+[dependencies]
+clap = { version = "3.0.14", features = ["derive"] }
+claim = { version = "0.3.1" }
+log = { version = "0.4" }
+env_logger = { version = "0.9.0" }
+glob = { version = "0.3" }
+serde = { version = "1.0", features = ["derive"] }
+serde_yaml = { version = "0.8" }
+reqwest = { version = "0.11", features = ["stream"] }
+futures-util = { version = "0.3.14" }
+indicatif = { version = "0.15.0" }
+tokio = { version = "1", features = ["full"] }
+tempfile = { version = "3.4.0" }
