+++ flake-ctl-v1/template/firecracker-flake.yaml
